Add files via upload
Overview
A linked list is a fundamental data structure that consists of a sequence of elements, where each element points to the next one in the sequence. In this implementation, we have created a simple singly linked list, which means each node in the list contains a reference to the next node, and the last node points to null.

Functionality
The linked list supports the following operations:

Add Last: Adds a new element to the end of the linked list.
Add First: Adds a new element to the beginning of the linked list.